pyproject.toml
[project]
name = "python-mcp-demos"
version = "0.1.0"
description = "Demonstration of Python FastMCP servers"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastmcp>=3.0.0b1",
    "debugpy>=1.8.0",
    "langchain-core>=0.3.0",
    "mcp>=1.3.0",
    "azure-identity>=1.25.1",
    "msgraph-sdk>=1.0.0",
    "dotenv-azd>=0.1.0",
    "langchain==1.0.0a5",
    "langchain-openai>=1.0.1",
    "langchain-mcp-adapters>=0.1.11",
    "azure-ai-agents>=1.1.0",
    "agent-framework-core==1.0.0rc5",
    "azure-cosmos>=4.9.0",
    "azure-monitor-opentelemetry>=1.8.3",
    "opentelemetry-instrumentation-starlette>=0.60b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.0",
    "logfire>=4.15.1",
    "azure-core-tracing-opentelemetry>=1.0.0b12"
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["D203"]
lint.isort.known-first-party = ["opentelemetry_middleware"]
